+++ bootmgr-rs-core/src/boot/bli.rs
@@ -9,7 +9,9 @@
 //! to use a tool such as `bootctl` to set the timeout, or set the next boot option.
 //!
 //! This module provides an essentially complete implementation of this interface as per the
-//! [Boot Loader Interface](https://systemd.io/BOOT_LOADER_INTERFACE/) specification.
+//! [Boot Loader Interface](https://systemd.io/BOOT_LOADER_INTERFACE/) specification. Only the features necessary
+//! for interaction with a tool like `bootctl` are implemented in this module. Even if `bootctl` defines more
+//! features, these are purely for reporting in `bootctl status`.
 
 use alloc::{format, string::ToString, vec::Vec};
 
@@ -42,6 +44,17 @@ const RANDOM_SEED_PATH: &CStr16 = cstr16!("\\loader\\random-seed");
 
 bitflags! {
     /// Feature flags for Boot Loader Interface.
+    ///
+    /// Note that half of these features aren't necessary for interacting with `bootctl`, they only exist to
+    /// specify the "features" that the bootloader supports in `bootctl status`.
+    ///
+    /// Some of these features are flagged as not supported, even if `bootmgr-rs` may support these features,
+    /// due to different paths that `bootmgr-rs` may use or because it uses a different configuration format
+    /// (such as `LOAD_DRIVER` and `SAVED_ENTRY`).
+    /// 
+    /// The only feature flags that are necessary for direct interaction with `bootctl` are `TIMEOUT`, 
+    /// `TIMEOUT_ONESHOT`, `ENTRY_DEFAULT`, `ENTRY_ONESHOT`, `BOOT_COUNTER`, `XBOOTLDR`, `RANDOM_SEED`,
+    /// and `MENU_DISABLED`.
     struct LoaderFeatures: u64 {
         const TIMEOUT = 1 << 0;
         const TIMEOUT_ONESHOT = 1 << 1;
@@ -50,7 +63,18 @@ bitflags! {
         const BOOT_COUNTER = 1 << 4;
         const XBOOTLDR = 1 << 5;
         const RANDOM_SEED = 1 << 6;
+        const LOAD_DRIVER = 1 << 7;
+        const SORT_KEY = 1 << 8;
+        const SAVED_ENTRY = 1 << 9;
+        const DEVICETREE = 1 << 10;
+        const SECUREBOOT_ENROLL = 1 << 11;
+        const RETAIN_SHIM = 1 << 12;
         const MENU_DISABLED = 1 << 13;
+        const MULTI_PROFILE_UKI = 1 << 14;
+        const REPORT_URL = 1 << 15;
+        const TYPE1_UKI = 1 << 16;
+        const TYPE1_UKI_URL = 1 << 17;
+        const TPM2_ACTIVE_PCR_BANKS = 1 << 18;
     }
 }
 
@@ -67,6 +91,9 @@ pub(crate) fn export_variables() -> BootResult<()> {
         | LoaderFeatures::BOOT_COUNTER
         | LoaderFeatures::XBOOTLDR
         | LoaderFeatures::RANDOM_SEED
+        | LoaderFeatures::SORT_KEY
+        | LoaderFeatures::DEVICETREE
+        | LoaderFeatures::RETAIN_SHIM
         | LoaderFeatures::MENU_DISABLED; // this is frontend dependent, depending on how input events are handled.
 
     let time = str_to_cstr(&timer_usec().to_string())?;
@@ -235,25 +262,19 @@ fn match_timeout(timeout: &uefi::CStr16) -> Option<i64> {
 /// This will generate a random seed as per the Boot Loader Interface, which list that the random seed should be hashed
 /// with the available source of entropy (the Rng protocol), the random seed, and the system token. Additionally, it will
 /// also hash in the available time (microseconds since boot). Entropy is gathered on a best-effort basis, which means
-/// that errors that may occur from the available sources are ignored. The only source which is required is the random seed,
-/// since this also indicates if the operating system supports using this random seed in the first place.
+/// that errors that may occur from the available sources are ignored.
 ///
 /// # Errors
 ///
-/// May return an `Error` if the filesystem could not be opened, the random seed could not be read, or the finalized
-/// seed could not be written back to.
+/// May return an `Error` if the filesystem could not be opened, or the finalized seed could not be written back to.
 pub(crate) fn generate_random_seed() -> BootResult<()> {
     let mut fs = UefiFileSystem::from_image_fs()?;
 
-    if !fs.exists(RANDOM_SEED_PATH) {
-        return Ok(()); // return immediately if there isnt even a random seed to begin with, before initializing hasher.
-    }
-
     let mut hasher = Sha256::new();
 
-    let content = fs.read(RANDOM_SEED_PATH)?;
-
-    hasher.update(&content);
+    if let Ok(content) = fs.read(RANDOM_SEED_PATH) {
+        hasher.update(&content);
+    }
 
     if let Ok((token, _)) = runtime::get_variable_boxed(cstr16!("LoaderSystemToken"), &BLI_VENDOR) {
         hasher.update(&token);
@@ -272,6 +293,7 @@ pub(crate) fn generate_random_seed() -> BootResult<()> {
     let result = hasher.finalize();
 
     let _ = fs.delete(RANDOM_SEED_PATH);
+    fs.create(RANDOM_SEED_PATH)?;
     fs.write(RANDOM_SEED_PATH, &result)?;
 
     Ok(())